WIP: convert issue food chain to generators
mypy is happy and unit tests have been fixed up so they pass again
(mostly by liberal application of `list()` to force execution of the
generator).

This still needs work to fix the logic where it is assumed that the
`issues` property is a list, namely where we try to take the length
of it or consume it more than once.