Unwind default console output
Previously we marshalled all issues into a single `echo()` call, which
resulted in buffering all output until the completion of the scan.

Now we use a loop (like "compact") and issue a separate `echo()` for
each issue, allowing early output to be produced prior to the generation
of later issues. This also gets rid of the unwanted "empty output for no
issues" behavior introduced by an earlier commit.