Hold issues for storage to files
Address the last remaining wart; make `echo_result()` buffer issues
(even as it streams them to console) and return the list so it's
possible to write all issues to files at the end if desired. This
returns `write_outputs()` to its original semantics, namely consumption
of a List rather than a Generator.

Arguably it might be nicer to create files as we go, but that would
constitute a behavior change (files not all done right at the end) and
need some code refactoring.